Versão previa v0.0.6
add public img

+++ app/Http/Controllers/ProductController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
     public function index(Request $request)
     {
+        if(!Auth::user()->is_buyer)
+            return redirect()->back()->with('error', 'Você não tem permissão para acessar produtos');
         $search = $request->get('search', '');
         $products = Product::where('description', 'like', "%{$search}%")
             ->orWhere('brand', 'like', "%{$search}%")
@@ -29,12 +32,16 @@ class ProductController extends Controller
 
     public function create(Request $request)
     {
+        if(!Auth::user()->is_buyer)
+            return redirect()->back()->with('error', 'Você não tem permissão para acessar produtos');
         $categories = Category::all();
         return view('products.create', compact('categories'));
     }
 
     public function store(Request $request)
     {
+        if(!Auth::user()->is_buyer)
+            return redirect()->back()->with('error', 'Você não tem permissão para acessar produtos');
         $previousUrl = $request->input('previous');
 
         $request->validate([
@@ -60,6 +67,11 @@ class ProductController extends Controller
     public function edit(string|int $hashedId)
     {
         $product = Product::find($this->decodeHash($hashedId));
+
+        if($product->price->count() > 0){
+            return redirect()->route('products.index')->with('error', 'Este produto está em um orçamento e não pode ser editado');
+        }
+
         $product->hashedId = $hashedId;
         $categories = Category::all();
         return view('products.show', compact('product', 'categories'));
@@ -74,7 +86,12 @@ class ProductController extends Controller
 
     public function destroy(string $hashedId)
     {
+        if(!Auth::user()->is_buyer)
+            return redirect()->back()->with('error', 'Você não tem permissão para excluir produtos');
         $product = Product::find($this->decodeHash($hashedId));
+        if($product->price->count() > 0){
+            return redirect()->route('products.index')->with('error', 'Este produto está em um orçamento e não pode ser excluido');
+        }
         $product->delete();
         return redirect()->route('products.index')->with('success', 'Product deleted successfully');
     }
